+++ Assets/Scripts/FirstPersonController.cs
@@ -310,7 +310,7 @@ public class FirstPersonController : MonoBehaviour
             ball.TryKick(gameObject);
     }
 
-    private const string GameVersion = "0.3.184-dev";
+    private const string GameVersion = "0.3.185-dev";
 
     private float lastSpeed;
     private bool lastSprinting;